N-len string comparison increased flexibility (#59)
* ~ preparatory work

* N-len string comparisons now work regardless of integral type of length argument

---------

+++ src/xtests.core.cpp
@@ -495,7 +495,7 @@ namespace
         ,   char const*             expr
         ,   wchar_t const*          expected
         ,   wchar_t const*          actual
-        ,   int                     n /* exact if +ve; limit if -ve */
+        ,   ptrdiff_t               n /* exact if +ve; limit if -ve */
         ,   size_t                  cchExpected
         ,   size_t                  cchActual
         ,   xtests_comparison_t     comp
@@ -1702,7 +1702,7 @@ xtests_testWideStringsN_(
 ,   char const*         expr
 ,   wchar_t const*      expected
 ,   wchar_t const*      actual
-,   int                 n /* exact if +ve; limit if -ve */
+,   ptrdiff_t           n /* exact if +ve; limit if -ve */
 ,   size_t              cchExpected
 ,   size_t              cchActual
 ,   xtests_comparison_t comp
@@ -4809,7 +4809,7 @@ RunnerInfo::TestWideStringsN(
 ,   char const*         expr
 ,   wchar_t const*      expected
 ,   wchar_t const*      actual
-,   int                 n
+,   ptrdiff_t           n
 ,   size_t              cchExpected
 ,   size_t              cchActual
 ,   xtests_comparison_t comp)
